+++ react/components/AddressRedeemForm.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 import { FormattedMessage } from 'react-intl'
 import Button from 'vtex.styleguide/Button'
 import ProfileField from '@vtex/profile-form/lib/ProfileField'
-import StyleguideInput from './StyleguideInput'
 
+import StyleguideInput from './StyleguideInput'
 import PhoneInputIcon from './PhoneInputIcon'
 import withImage from './withImage'
 
@@ -81,7 +81,7 @@ class AddressRedeemForm extends Component {
     const profilePhone = profile[homePhoneField.name]
 
     return (
-      <form className="vtex-address-locator__address-redeem w-100 pv7 ph6" onSubmit={this.handleSubmit}>
+      <form className="vtex-address-locator__address-redeem w-100 pv7 ph6 br2 bg-white" onSubmit={this.handleSubmit}>
         <div className="mb5 relative input--icon-left">
           <ProfileField
             key={homePhoneField.name}
@@ -91,6 +91,8 @@ class AddressRedeemForm extends Component {
             options={{
               placeholder: '(99) 99999-9999',
               size: 'large',
+              inputmode: "numeric",
+              type: "tel"
             }}
             Input={StyleguideInput}
           />