+++ react/components/withImage.js
@@ -22,15 +22,14 @@ export default function withImage(getImageFilename) {
       componentDidMount() {
         const imageName = getImageFilename(this.props)
         import(`../images/${imageName}`).then(imageSrc => {
-          this.setState({
-            imageSrc: imageSrc.default,
-          })
+          this.setState({ imageSrc: imageSrc.default })
         })
       }
 
       render() {
         const { imageSrc } = this.state
-        return <WrappedComponent {...this.props} imageSrc={imageSrc} />
+        const src = imageSrc && imageSrc.default || imageSrc
+        return <WrappedComponent {...this.props} imageSrc={src} />
       }
     }