+++ react/index.js
@@ -4,10 +4,10 @@ import { Adopt } from 'react-adopt'
 import { compose } from 'recompose'
 import { withRuntimeContext } from 'render'
 import PropTypes from 'prop-types'
-
-import Tabs from 'vtex.styleguide/Tabs'
-import Tab from 'vtex.styleguide/Tab'
 import { orderFormConsumer, contextPropTypes } from 'vtex.store/OrderFormContext'
+import { Tab, Tabs, Spinner } from 'vtex.styleguide'
+
+import logisticsQuery from './queries/logistics.gql'
 import AddressSearch from './components/AddressSearch'
 import AddressRedeem from './components/AddressRedeem'
 import './global.css'
@@ -66,50 +66,62 @@ class AddressLocator extends Component {
 
   render() {
     const { currentTab } = this.state
-    const { orderFormContext } = this.props
+    const { orderFormContext, hideTabs, hideTitle } = this.props
+
+    const addressSearchTab = <FormattedMessage id="address-locator.address-search-tab" />
+    const addressRedeemTab = <FormattedMessage id="address-locator.address-redeem-tab" />
+    const addressSearch = (
+      <Query query={logisticsQuery}>
+        {({ data, loading }) => {
+          if (loading) {
+            return <Spinner />
+          }
+
+          const { googleMapsKey } = data.logistics
+
+          return (
+            <AddressSearch
+              googleMapURL={`https://maps.googleapis.com/maps/api/js?key=${googleMapsKey}&v=3.exp&libraries=places`}
+              googleMapKey={googleMapsKey}
+              loadingElement={<div className="h-100" />}
+              orderFormContext={orderFormContext}
+              onOrderFormUpdated={this.handleOrderFormUpdated}
+            />
+          )
+        }}
+      </Query>
+    )
+
+    const tabs = (
+      <Tabs fullWidth>
+        <Tab
+          label={addressSearchTab}
+          active={currentTab === 1}
+          onClick={() => this.handleTabChange(1)}
+        >
+          {addressSearch}
+        </Tab>
+        <Tab
+          label={addressRedeemTab}
+          active={currentTab === 2}
+          onClick={() => this.handleTabChange(2)}
+        >
+          <AddressRedeem
+            orderFormContext={orderFormContext}
+            onOrderFormUpdated={this.handleOrderFormUpdated}
+          />
+        </Tab>
+      </Tabs>
+    )
 
     return (
-      <div>
-        <Adopt
-          mapper={{
-            title: <FormattedMessage id="address-locator.order-title" />,
-          }}
-        >
-          {({ title }) => <span className="db b f1 pa5 tl tc-m">{title}</span>}
-        </Adopt>
-        <div className="vtex-address-locator w-100 w-60-m w-40-l center flex flex-column justify-center items-center pa5">
-          <Adopt
-            mapper={{
-              addressSearchTab: <FormattedMessage id="address-locator.address-search-tab" />,
-              addressRedeemTab: <FormattedMessage id="address-locator.address-redeem-tab" />,
-            }}
-          >
-            {({ addressSearchTab, addressRedeemTab }) => (
-              <Tabs fullWidth>
-                <Tab
-                  label={addressSearchTab}
-                  active={currentTab === 1}
-                  onClick={() => this.handleTabChange(1)}
-                >
-                  <AddressSearch
-                    orderFormContext={orderFormContext}
-                    onOrderFormUpdated={this.handleOrderFormUpdated}
-                  />
-                </Tab>
-                <Tab
-                  label={addressRedeemTab}
-                  active={currentTab === 2}
-                  onClick={() => this.handleTabChange(2)}
-                >
-                  <AddressRedeem
-                    orderFormContext={orderFormContext}
-                    onOrderFormUpdated={this.handleOrderFormUpdated}
-                  />
-                </Tab>
-              </Tabs>
-            )}
-          </Adopt>
-        </div>
+      <div className="vtex-address-locator w-100 w-50-m center flex flex-column justify-center items-center pa6">
+        {!hideTitle && (
+          <h1 className="db b f1 mb7 pa5 tl tc-m">
+            <FormattedMessage id="address-locator.order-title" />
+          </h1>
+        )}
+        {hideTabs ? addressSearch : tabs}
       </div>
     )
   }