+++ CHANGELOG.md
@@ -9,4 +9,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - MVP of `Address Locator` app.
 - Added component to locate address and insert data into orderform
-- Component to format and validate phone number
+- Redeem addresses by phone number